requirements.txt
slackweb
python-dotenv